Initial commit to archive programs done in 2018 and 2014
The program is declined in two versions, one at the plot scale and one at the landscape scale. 
The plot-scale program is "Modele_paysage" and is based on Gotheron's zero project, a circular orchard integrating a diversity of crops to fight detrimental species and attract beneficials ones. It was developed in 2017 and 2018, and is based on 2014's program.
The landscape-scale program is "Modele_Daniele", it was developed in 2014 and illustrates the same dynamics with the influence of habitat amount in the landscape and pesticide use.